AyurvedaRAG/requirements.txt
fastapi
inngest
python-dotenv
qdrant-client
openai
pypdf
uvicorn
requests
fpdf2
